requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
pywin32>=306
pyperclip>=1.8.0
openai>=2.9.0
ttkbootstrap>=1.10.1
